restructure test code
